pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rekit"
version = "0.1.0"
description = "Reverse Engineering Toolkit for Mobile APIs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Baris Erdem" }]
keywords = ["reverse-engineering", "api", "mobile", "mitm", "apk", "tls"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "jinja2>=3.0",
    "requests>=2.28",
]

[project.optional-dependencies]
tls = ["curl_cffi>=0.5"]
frida = ["frida-tools>=12.0"]
all = ["curl_cffi>=0.5", "frida-tools>=12.0"]

[project.scripts]
rekit = "rekit.cli:app"

[project.urls]
Homepage = "https://github.com/b-erdem/rekit"
Repository = "https://github.com/b-erdem/rekit"
Issues = "https://github.com/b-erdem/rekit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/rekit"]
